Extract query rewriting tools to library
